+++ daeng/src/components/preference/PreferenceForm.jsx
@@ -1,11 +1,19 @@
 import PropTypes from "prop-types";
-import styled from "styled-components";
 import ConfirmBtn from "../commons/ConfirmBtn";
 import PreferencePlaceOption from "../commons/PreferencePlaceOption";
 import PreferenceFavoriteOption from "../commons/PreferenceFavoriteOption";
 import Loading from "../commons/Loading";
 import AlertDialog from "../commons/SweetAlert";
 import { placeFeatures, placeTypes } from "../../data/CommonCode";
+import {
+  Wrap,
+  Section,
+  Title,
+  StyledParagraph,
+  OptionContainer,
+  StyledParagraph2,
+  Footer,
+} from "./PreferenceForm.styles";
 
 function PreferenceForm({
   selectedPlaceOptions,
@@ -100,83 +108,5 @@ PreferenceForm.defaultProps = {
   isLoading: false,
   confirmButtonLabel: "확인",
 };
-
-const Wrap = styled.div`
-  padding: 20px;
-  @media (max-width: 554px) {
-    padding-left: 20px;
-    padding-right: 20px;
-    padding-top: 0px;
-  }
-`;
-
-const Section = styled.div`
-  margin-bottom: 40px;
-`;
-
-const Title = styled.h3`
-  text-align: left;
-  margin-left: 10px;
-  font-size: 17px;
-  margin-bottom: 10px;
-
-  @media (max-width: 554px) {
-    font-size: 14px;
-    margin-left: 20px;
-  }
-`;
-
-const StyledParagraph = styled.p`
-  font-size: 12px;
-  color: #ff69a9;
-  font-weight: bold;
-  margin-top: 9px;
-  margin-left: 10px;
-  display: flex;
-
-  @media (max-width: 554px) {
-    font-size: 10px;
-    margin-left: 20px;
-  }
-`;
-
-const OptionContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  margin-top: 10px;
-  gap: 5px;
-  margin-left: 5px;
-
-  @media (max-width: 554px) {
-    gap: 3px;
-  }
-`;
-
-const StyledParagraph2 = styled.p`
-  font-size: 14px;
-  color: red;
-  font-weight: bold;
-  margin-top: 80px;
-  margin-bottom: 20px;
-  display: flex;
-  justify-content: center;
-
-  @media (max-width: 554px) {
-    font-size: 10px;
-    margin-left: 10px;
-    text-align: center;
-    margin-top: 60px;
-  }
-`;
-
-const Footer = styled.div`
-  margin-left: 15px;
-  padding-bottom: 10px;
-
-  @media (max-width: 554px) {
-    margin-left: 20px;
-  }
-`;
 
 export default PreferenceForm;
