+++ daeng/src/components/user/UserForm.styles.jsx
@@ -1,0 +1,127 @@
+import styled from "styled-components";
+
+export const UserContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 3%;
+  margin-left: 4%;
+`;
+
+export const InputEmailContainer = styled.div`
+  display: flex;
+  align-items: center;
+  background-color: rgba(228, 228, 228, 0.2);
+  width: 95%;
+  height: 44px;
+  border-radius: 5px;
+  border: 0.5px solid #E4E4E4;
+  color: black;
+  padding: 0 10px;
+  margin-bottom: 20px;
+`;
+
+export const InputBox = styled.div`
+  display: flex;
+  align-items: center;
+  width: 95%;
+  height: 44px;
+  border-radius: 5px;
+  border: 0.5px solid #E4E4E4;
+  color: black;
+  padding: 0 10px;
+  margin-bottom: 10px;
+`;
+
+export const Input = styled.input`
+  flex: 1;
+  border: none;
+  background: transparent;
+  outline: none;
+  font-size: 14px;
+  color: black;
+
+  ::placeholder {
+    color: #b3b3b3;
+  }
+
+  &:disabled {
+    color: #b3b3b3;
+  }
+
+  @media (max-width: 554px) {
+    font-size: 10px;
+  }
+`;
+
+export const Icon = styled.img`
+  width: 20px;
+  cursor: pointer;
+  margin-right: 10px;
+`;
+
+export const DuplicateBtn = styled.button`
+  width: 15%;
+  height: 23px;
+  border-radius: 10px;
+  border: none;
+  font-size: 11px;
+  cursor: pointer;
+  background-color: #ff69a9;
+  color: white;
+
+  &:hover {
+    background-color: #f9a9d4;
+  }
+
+  @media (max-width: 554px) {
+    font-size: 8px;
+  }
+`;
+
+export const InputAlert = styled.p`
+  color: #ff69a9;
+  font-size: 12px;
+  display: flex;
+  margin-top: -1px;
+  flex-direction: flex-start;
+  margin-bottom: 4%;
+
+  @media (max-width: 554px) {
+    font-size: 10px;
+  }
+`;
+
+export const SelectionContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  gap: 10px;
+`;
+
+export const SelectBox = styled.select`
+  flex: 1;
+  height: 44px;
+  border-radius: 5px;
+  padding: 0;
+  margin-right: 18px;
+  margin-bottom: 10px;
+  border: 0.5px solid #e4e4e4;
+  font-size: 14px;
+  text-align-last: center;
+  background-color: white;
+  color: black;
+  line-height: 44px;
+
+  &:focus {
+    border-color: #ff69a9;
+    outline: none;
+  }
+
+  @media (max-width: 554px) {
+    font-size: 11px;
+  }
+`;
+
+export const ConfirmContainer = styled.div`
+  display: flex;
+  margin-top: 40px;
+`;
